+++ llvm/lib/Target/X86/X86TargetTransformInfo.h
@@ -348,8 +348,8 @@ private:
                                   Type *DataTy, const Value *Ptr,
                                   Align Alignment, unsigned AddressSpace) const;
 
-  int getGatherOverhead() const;
-  int getScatterOverhead() const;
+  int getGatherOverhead(Type *SrcVTy) const;
+  int getScatterOverhead(Type *SrcVTy) const;
 
   /// @}
 };